Enhancements to NSIS Installer
Set the NSIS installer to allow users to install outside their profile directory (per-user). Add fancy graphics for the install header and sidebar. Support HiDPI screens in the installer so it doesn’t look so blurry. Remove the external electron-builder.yml as build config is being used in package.json.